+++ resources/views/checkout/show.blade.php
@@ -1,0 +1,3 @@
+<x-layouts::storefront :title="__('Checkout')">
+    <livewire:storefront.checkout-page />
+</x-layouts::storefront>